pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "transport-atlas"
version = "0.1.0"
description = "A browsable atlas of transportation research papers (TR A-F, IEEE T-ITS/T-IV/ITS Mag/IV Symp)"
requires-python = ">=3.10"
authors = [{name = "Seongjin Choi", email = "chois@umn.edu"}]
dependencies = [
    "requests>=2.31",
    "pyyaml>=6.0",
    "pandas>=2.2",
    "duckdb>=1.0",
    "pyarrow>=15",
    "networkx>=3.2",
    "python-igraph>=0.11",
    "leidenalg>=0.10",
    "fa2-modified>=0.3.10",
    "jinja2>=3.1",
    "tenacity>=8.2",
    "tqdm>=4.66",
    "rapidfuzz>=3.6",
    "unidecode>=1.3",
    "python-dotenv>=1.0",
    "lxml>=5.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.4"]
topics = ["bertopic>=0.16", "sentence-transformers>=2.7", "umap-learn>=0.5"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
